+++ sim/agent/testagent/AntGoHomeAntTest.hpp
@@ -17,7 +17,7 @@ namespace cdc
 		AntGoHomeAntTest(GuiEventManager& manager, AntHome& home, NavGraphHelper& navGraphHelper, const Node& startNode);
 		virtual ~AntGoHomeAntTest();
 
-		virtual void update(long ticks, const Percept& percept) override;
+		virtual void update(uint ticks, const Percept& percept) override;
 
 		virtual bool isGoalFinished() const override;
