+++ sim/agent/testagent/AntMoveToNodeAntTest.hpp
@@ -17,7 +17,7 @@ namespace cdc
 		AntMoveToNodeAntTest(GuiEventManager& manager, AntHome& home, NavGraphHelper& navGraphHelper, const Node& startNode, Node& target);
 		virtual ~AntMoveToNodeAntTest();
 
-		virtual void update(long ticks, const Percept& percept) override;
+		virtual void update(uint ticks, const Percept& percept) override;
 
 		virtual bool isGoalFinished() const override;
